Scene Tweaking
This fixes an issue where we were creating the scenes all up front to
now be created on demand. There was also an issue that needed to be
fixed where the form elements were being created PRIOR to the scenes,
this change ensures that the form fields are PART of the menu. The issue
I saw previously is that the buttons would lose their text when
switching scenes.